requirements.txt
requests
beautifulsoup4
discord.py>=1.3.1
python-dotenv
pytz